requirements.txt
pytest
pytest-playwright 